Move new integration tests to existing module
- Moved the test class
- Deactivated all old tests
- Changed how docker-entrypoint.sh resource is converted to file
  Before it went via URI, which doesn't work when executing tests
  from a different module. Write resource to temp file instead,
