+++ integration-tests/src/test/java/integration/roles/RolesResourceTest.java
@@ -18,10 +18,12 @@ package integration.roles;
 
 import integration.BaseRestTest;
 import integration.RequiresAuthentication;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import static com.jayway.restassured.RestAssured.given;
 
+@Ignore("legacy test that should be converted or deleted")
 @RequiresAuthentication
 public class RolesResourceTest extends BaseRestTest {
